backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
google-generativeai==0.8.3
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
anthropic==0.39.0
